修复了一些在微信互动页面上（canJoin，joined等）涉及基于微信定位的自主签到的bug
已完成：
（3）微信端开辟一个入口“地理信息”，用来采集经纬度数据信息，用来设置活动室和地理点位的经纬度坐标；
待完成：
（6）为活动增加"补签"功能，用来扫描活动参与者的个人二维码完成签到（即便活动已经过期/该用户没有报名该活动等，也一次性完成签到/签退/积分/积时），也可用来对没有微信的志愿者完成签到积分功能。

+++ src/cc/natapp4/ddaig/action/PersonalCenterAction.java
@@ -124,19 +124,19 @@ public class PersonalCenterAction extends ActionSupport {
 	}
 	// 基于位置的签到/签退，JS-SDK或获取签到/签退者的地理位置坐标，并以AJAX的post请求参数的形式传递过来，下面就是记录经度和纬度的属性驱动
 	// 纬度
-	private long latitude;
-	public long getLatitude() {
+	private double latitude;
+	public double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 	// 经度
-	private long longitude;
-	public long getLongitude() {
+	private double longitude;
+	public double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(long longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 	/*
